Another fix for Bart.
svn path=/branches/koffice/1.5/koffice/; revision=523650

+++ krita/core/tiles/kis_tilemanager.cc
@@ -278,7 +278,7 @@ void KisTileManager::toSwap(TileInfo* info) {
 
         if (data == 0) { // No position found or free, create a new
             long pagesize = sysconf(_SC_PAGESIZE);
-            int newsize = m_fileSize + info->size;
+            off_t newsize = m_fileSize + info->size;
             newsize = newsize + newsize % pagesize;
 
             if (ftruncate(m_tempFile.handle(), newsize)) {
